Modifications for vspace partitioning in progress. Logic to check
Modified until line 160 in ./DG_lin_convec.f90, following the program
flow. Next, INICON* subroutines must be modified.
A scheme of the communicators showing the meaning of all MPI-related
integers  must be added to the repo, in the 'doc' directory.

 The old master-slave pattern is replicated NVSPACEPART=MPI_SIZE_V times,
with MPI_SIZE_V master ranks doing the same and sending data to the
ranks in their communicators (the masters do the setup)
This is suboptimal but easy to implement, hopefully.

Removing the master should not be too difficult - see issue #2, but
gains should be less significant.